requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
binance==0.3
certifi==2022.6.15
charset-normalizer==2.0.12
CurrencyConverter==0.16.12
dateparser==1.1.1
dj-database-url==0.5.0
Django==4.0.2
django-environ==0.8.1
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
psycopg2==2.9.3
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
validate-email==1.3
websockets==10.3
whitenoise==6.0.0
yarl==1.7.2
